+++ src/main/java/com/binance/client/model/event/SymbolTickerEvent.java
@@ -9,7 +9,7 @@ public class SymbolTickerEvent {
 
     private String eventType;
 
-    private Integer eventTime;
+    private Long eventTime;
 
     private String symbol;
 
@@ -43,15 +43,15 @@ public class SymbolTickerEvent {
 
     private BigDecimal totalTradedQuoteAssetVolume;
 
-    private Integer openTime;
+    private Long openTime;
 
-    private Integer closeTime;
+    private Long closeTime;
 
-    private Integer firstId;
+    private Long firstId;
 
-    private Integer lastId;
+    private Long lastId;
 
-    private Integer count;
+    private Long count;
 
     public String getEventType() {
         return eventType;
@@ -61,11 +61,11 @@ public class SymbolTickerEvent {
         this.eventType = eventType;
     }
 
-    public Integer getEventTime() {
+    public Long getEventTime() {
         return eventTime;
     }
 
-    public void setEventTime(Integer eventTime) {
+    public void setEventTime(Long eventTime) {
         this.eventTime = eventTime;
     }
 
@@ -197,43 +197,43 @@ public class SymbolTickerEvent {
         this.totalTradedQuoteAssetVolume = totalTradedQuoteAssetVolume;
     }
 
-    public Integer getOpenTime() {
+    public Long getOpenTime() {
         return openTime;
     }
 
-    public void setOpenTime(Integer openTime) {
+    public void setOpenTime(Long openTime) {
         this.openTime = openTime;
     }
 
-    public Integer getCloseTime() {
+    public Long getCloseTime() {
         return closeTime;
     }
 
-    public void setCloseTime(Integer closeTime) {
+    public void setCloseTime(Long closeTime) {
         this.closeTime = closeTime;
     }
 
-    public Integer getFirstId() {
+    public Long getFirstId() {
         return firstId;
     }
 
-    public void setFirstId(Integer firstId) {
+    public void setFirstId(Long firstId) {
         this.firstId = firstId;
     }
 
-    public Integer getLastId() {
+    public Long getLastId() {
         return lastId;
     }
 
-    public void setLastId(Integer lastId) {
+    public void setLastId(Long lastId) {
         this.lastId = lastId;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(Long count) {
         this.count = count;
     }
